ProFile/telegram_bot/requirements.txt

aiogram==3.10.0
SQLAlchemy==2.0.34
aiosqlite==0.20.0
python-dotenv==1.0.1
pydantic==2.9.2
